automation for the delete participants feature
