+++ study/test/src/org/labkey/test/tests/study/StudyManageParticipantsTest.java
@@ -1,0 +1,344 @@
+/*
+ * Copyright (c) 2016-2019 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.labkey.test.tests.study;
+
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+import org.labkey.remoteapi.Connection;
+import org.labkey.remoteapi.domain.CreateDomainCommand;
+import org.labkey.remoteapi.query.InsertRowsCommand;
+import org.labkey.test.BaseWebDriverTest;
+import org.labkey.test.Locator;
+import org.labkey.test.TestFileUtils;
+import org.labkey.test.WebTestHelper;
+import org.labkey.test.categories.Daily;
+import org.labkey.test.components.ext4.Window;
+import org.labkey.test.pages.study.DatasetDesignerPage;
+import org.labkey.test.params.FieldDefinition;
+import org.labkey.test.util.Ext4Helper;
+import org.labkey.test.util.Maps;
+import org.labkey.test.util.StudyHelper;
+
+import java.io.File;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Random;
+import java.util.stream.Collectors;
+
+import static org.junit.Assert.assertFalse;
+
+@Category(Daily.class)
+@BaseWebDriverTest.ClassTimeout(minutes = 6)
+public class StudyManageParticipantsTest extends BaseWebDriverTest
+{
+    private static final String PROJECT_NAME =  "Merge Participants Test Project";
+    private static final String STUDY_NAME = "Merge Participants Test Study";
+    private static final String FOLDER_NAME =  "Merge Participants Test Folder";
+    private static final String DELETE_PARTICIPANTS_FOLDER_NAME = "Delete Participants Test Folder";
+    private static final File STUDY_ZIP = TestFileUtils.getSampleData("studies/LabkeyDemoStudy.zip");
+    private static final String SUBJECT_NOUN = "Participant";
+    private static final String SUBJECT_NOUN_PLURAL = "Participants";
+    private static final String ALIAS_DATASET = "MyAliasDataset";
+    private static final String ALIAS_COLUMN = "MyAlias";
+    private static final String SOURCE_COLUMN = "MySource";
+
+    private static final String PTID_WITH_ALIAS = "249318596";
+    private static final String PTID_NO_ALIAS = "249320107";
+    private static final String PTID_NEW_1 = "xyz987";
+    private static final String PTID_NEW_2 = "249325717";
+
+    private static final String ALIAS_SOURCE_1 = "a";
+    private static final String ALIAS_SOURCE_2 = "b";
+    private static final Locator.XPathLocator CREATE_ALIAS_CB = Ext4Helper.Locators.ext4CheckboxById("createAliasCB");
+    private static final Locator OLD_ID_FIELD = Locator.name("oldIdField-inputEl");
+    private static final Locator NEW_ID_FIELD = Locator.name("newIdField-inputEl");
+    private static final Locator ALIAS_SOURCE_FIELD = Locator.name("aliasSourceField-inputEl");
+
+    private static final int MERGE_SUCCESS_TIMEOUT = 60000;
+
+    @Override
+    protected String getProjectName()
+    {
+        return PROJECT_NAME;
+    }
+
+    protected String getStudyLabel()
+    {
+        return STUDY_NAME;
+    }
+
+    protected String getFolderName()
+    {
+        return FOLDER_NAME;
+    }
+
+    @BeforeClass
+    public static void setupProject() throws Exception
+    {
+        StudyManageParticipantsTest test = getCurrentTest();
+        test.doCreateSteps();
+    }
+
+    protected void doCreateSteps()
+    {
+        _containerHelper.createProject(getProjectName(), null);
+        _containerHelper.createSubfolder(getProjectName(), getFolderName(), "Study");
+
+        clickFolder(getFolderName());
+        log("Import LabkeyDemoStudy");
+        importStudyFromZip(STUDY_ZIP);
+
+        log("Creating study to test deleting participants");
+        _containerHelper.createSubfolder(getProjectName(), DELETE_PARTICIPANTS_FOLDER_NAME, "Study");
+        _studyHelper.startCreateStudy()
+                .setTimepointType(StudyHelper.TimepointType.VISIT)
+                .createStudy();
+    }
+
+    @Test
+    public void testDeleteParticipants() throws Exception
+    {
+        // set up some datasets and participant IDs
+        log("Creating visit study datasets");
+
+        final String containerPath = String.format("%s/%s", getProjectName(), DELETE_PARTICIPANTS_FOLDER_NAME);
+        Connection conn = createDefaultConnection();
+
+        List<String> datasets = List.of(
+                "Demographics",
+                "Physical_Exam",
+                "Infusion-1",
+                "Infusion_2"
+        );
+
+        List<String> ptids = List.of(
+                "123_abc-345",
+                "123_def-345",
+                "123_ghi-345",
+                "456_abc-345",
+                "789_abc-345"
+        );
+
+        List<String> participantsToDelete = List.of(
+                "123_abc-345",
+                "456_abc-345",
+                "789_abc-345"
+        );
+
+        log("Creating datasets and populating with some data");
+        Random rand = new Random();
+        for (String dataset : datasets)
+        {
+            createDataset(conn, "StudyDatasetVisit", dataset, containerPath);
+
+            // populate with some rows
+            List<Map<String, Object>> rows = ptids.stream().<Map<String, Object>>map(ptid -> Map.of(
+                    "Ptid", ptid,
+                    "Visit", 1,
+                    "IntField", rand.nextInt(),
+                    "DoubleField", rand.nextDouble()
+            )).collect(Collectors.toList());
+
+            InsertRowsCommand cmd = new InsertRowsCommand("Study", dataset);
+            cmd.setRows(rows);
+            cmd.execute(conn, containerPath);
+        }
+
+        goToProjectHome(getProjectName());
+        clickFolder(DELETE_PARTICIPANTS_FOLDER_NAME);
+        goToManageParticipants();
+
+        // delete participants through the UI
+        for (String ptid : participantsToDelete)
+        {
+            log(String.format("Deleting participant %s", ptid));
+            deleteParticipant(ptid);
+        }
+
+        // run the purge participant maintenance task before we verify the participants have been successfully deleted
+        startSystemMaintenance("PurgeStudyParticipants");
+        waitForSystemMaintenanceCompletion();
+
+        goToProjectHome(getProjectName());
+        clickFolder(DELETE_PARTICIPANTS_FOLDER_NAME);
+        goToManageParticipants();
+
+        Locator.XPathLocator ptidCombo = Locator.XPathLocator.tagWithClass("table", "participant-combo");
+        // validate that the deleted ptid no longer shows up
+        List<String> currentPtids = _ext4Helper.getComboBoxOptions(ptidCombo);
+        for (String ptid : participantsToDelete)
+            assertFalse(String.format("Participant %s was not successfully deleted", ptid), currentPtids.contains(ptid));
+    }
+
+    private void createDataset(Connection conn, String kindName, String datasetName, String folderName) throws Exception
+    {
+        CreateDomainCommand cmd = new CreateDomainCommand(kindName, datasetName);
+
+        cmd.getDomainDesign().setFields(List.of(
+                new FieldDefinition("Name", FieldDefinition.ColumnType.String),
+                new FieldDefinition("IntField", FieldDefinition.ColumnType.Integer),
+                new FieldDefinition("DoubleField", FieldDefinition.ColumnType.Double)
+        ));
+        if ("Demographics".equalsIgnoreCase(datasetName))
+            cmd.setOptions(Map.of("demographics", true));
+        cmd.execute(conn, folderName);
+    }
+
+    private void deleteParticipant(String participantId)
+    {
+        Locator.XPathLocator ptidCombo = Locator.XPathLocator.tagWithClass("table", "participant-combo");
+        _ext4Helper.selectComboBoxItem(ptidCombo, participantId);
+        waitAndClick(Ext4Helper.Locators.ext4Button("Delete"));
+        Window<?> confirm = Window.Window(getDriver()).withTitleContaining("Confirmation").waitFor();
+        confirm.clickButton("Yes", WAIT_FOR_EXT_MASK_TO_DISSAPEAR);
+
+        Window<?> success = Window.Window(getDriver()).withTitleContaining("Success").waitFor();
+        success.clickButton("OK", 0);
+    }
+
+    @Test
+    public void testMergeParticipants()
+    {
+        goToProjectHome(getProjectName());
+        clickFolder(FOLDER_NAME);
+        goToMergeParticipants();
+        log("Ensure alias creation controls absent when no alias dataset configured");
+        assertElementNotVisible(CREATE_ALIAS_CB);
+
+        configureAliases();
+        goToMergeParticipants();
+
+        log("Check have error for alias creation, as source field is blank");
+        assertElementPresent(ALIAS_SOURCE_FIELD); // if it's missing, the participant alias setup didn't work.
+        setFormElement(OLD_ID_FIELD, PTID_WITH_ALIAS);
+        setFormElement(NEW_ID_FIELD, PTID_NEW_1);
+        waitAndClick(Ext4Helper.Locators.ext4ButtonEnabled("Preview"));
+        waitForElement(Locator.tag("span").containing("Specimen data is not editable"), MERGE_SUCCESS_TIMEOUT);
+        // Error on missing value for source field.
+        assertElementPresent(Locator.tag("span").containing("Missing value for required property"));
+
+        log("Check not reporting conflict when no conflict exists, and warning on existing alias");
+        setFormElement(ALIAS_SOURCE_FIELD, ALIAS_SOURCE_2);
+        waitAndClick(Ext4Helper.Locators.ext4ButtonEnabled("Preview"));
+        waitForElement(Locator.tag("span").containing("Preview Complete"), MERGE_SUCCESS_TIMEOUT);
+        assertElementPresent(Locator.tag("td").containing("Warning: Specimen data is not editable"));
+        assertElementNotPresent(Locator.linkContainingText("Conflict!"));
+        assertElementPresent(Locator.tag("td").containing("Aliases are not updated by this process"));
+        assertElementPresent(Locator.tag("td").containing("Warning: " + PTID_WITH_ALIAS + " has existing aliases"));
+
+        log("Check merge successful and alias created.");
+        clickButton("Merge", 0);
+        waitForElement(Locator.tag("span").containing("Successfully merged"), MERGE_SUCCESS_TIMEOUT);
+        clickTab(SUBJECT_NOUN_PLURAL);
+        waitForElement(Locator.linkContainingText(PTID_NEW_1)); // New participantId now exists
+        assertElementNotPresent(Locator.linkContainingText(PTID_WITH_ALIAS));
+        click(Locator.linkContainingText(PTID_NEW_1));
+        waitForElement(Locator.tag("p").containing("b: " + PTID_WITH_ALIAS));
+
+        log("Check conflicts are correctly detected and prevent processing.");
+        goToMergeParticipants();
+        click(CREATE_ALIAS_CB);
+        setFormElement(OLD_ID_FIELD, PTID_NO_ALIAS);
+        setFormElement(NEW_ID_FIELD, PTID_NEW_2);
+        waitAndClick(Ext4Helper.Locators.ext4ButtonEnabled("Preview"));
+        waitForElement(Locator.tag("span").containing("Preview Complete"), MERGE_SUCCESS_TIMEOUT);
+        assertElementNotPresent(Locator.linkContainingText(PTID_NO_ALIAS + " has existing aliases"));
+        assertElementPresent(Locator.linkContainingText("Conflict!"));
+        clickButton("Merge", 0);
+        waitForElement(Locator.tag("span").containing("You must choose"));
+
+        log("Check url's to are correctly constructed");
+        final String url = WebTestHelper.buildRelativeUrl("study", PROJECT_NAME + "/" + FOLDER_NAME, "dataset",
+            Maps.of("datasetId", 5018, "Dataset.ParticipantId~in", PTID_NO_ALIAS + ";" + PTID_NEW_2))
+            .replace("%7E", "~"); // '~' isn't encoded in href
+        assertElementPresent(Locator.linkWithHref(url));
+
+        log("Resolve conflicts and check for correct row retention");
+        click(Locator.radioButtonByNameAndValue("conflict_Demographics", "new"));
+        click(Locator.radioButtonByNameAndValue("conflict_Participation and Genetic Consent", "old"));
+
+        clickButton("Merge", 0);
+        waitForElement(Locator.tag("span").containing("Successfully merged"), MERGE_SUCCESS_TIMEOUT);
+        clickTab(SUBJECT_NOUN_PLURAL);
+        waitAndClick(Locator.linkWithText(PTID_NEW_2));
+        waitForElement(Locator.tag("th").containing("5008: Demographics"));
+        assertElementPresent(Locator.tag("td").containing("Start Date").followingSibling("td").withText("2008-04-27")); //value from oldId demographic row
+        assertElementPresent(Locator.tag("td").containing("Genetic Consent").followingSibling("td").withText("false")); //value from newId consent row
+    }
+
+    private void configureAliases()
+    {
+        // Create alias dataset and insert some data
+        DatasetDesignerPage editDatasetPage = _studyHelper.goToManageDatasets()
+            .clickCreateNewDataset()
+            .setName(ALIAS_DATASET);
+        editDatasetPage.getFieldsPanel()
+            .manuallyDefineFields("Property");
+        editDatasetPage.getFieldsPanel()
+            .addField(ALIAS_COLUMN).setRequiredField(true);
+        editDatasetPage.getFieldsPanel()
+            .addField(SOURCE_COLUMN).setRequiredField(true);
+
+        editDatasetPage
+            .clickSave()
+            .clickViewData()
+            .getDataRegion()
+            .clickImportBulkData();
+        setFormElement(Locator.name("text"), "participantId\tdate\t" + ALIAS_COLUMN + "\t" + SOURCE_COLUMN + "\n" + PTID_WITH_ALIAS + "\t1/3/2014\tabc123\t" + ALIAS_SOURCE_1);
+        clickButton("Submit", "Dataset: " + ALIAS_DATASET);
+
+        // Configure new dataset as the alias dataset
+        goToManageParticipants();
+        _ext4Helper.selectComboBoxItem("Dataset Containing Aliases", ALIAS_DATASET);
+        _ext4Helper.selectComboBoxItem("Alias Column", ALIAS_COLUMN);
+        _ext4Helper.selectComboBoxItem("Source Column", SOURCE_COLUMN);
+        clickButton("Save Changes", 0);
+        waitForText(SUBJECT_NOUN + " alias settings saved successfully");
+        clickButton("OK", 0);
+
+        // Refresh the page and make sure the configuration stuck
+        refresh();
+        waitForFormElementToEqual(Locator.input("datasetCombo"), ALIAS_DATASET);
+
+        clickButton("Done");
+    }
+
+    @Override
+    protected BrowserType bestBrowser()
+    {
+        return BrowserType.CHROME;
+    }
+
+    private void goToManageParticipants()
+    {
+        goToManageStudy();
+        clickAndWait(Locator.linkContainingText("Manage Participants"));
+    }
+
+    private void goToMergeParticipants()
+    {
+        goToManageParticipants();
+        clickButton("Change or Merge " + SUBJECT_NOUN + " IDs");
+    }
+
+    @Override
+    public List<String> getAssociatedModules()
+    {
+        return Arrays.asList("study");
+    }
+}
